gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = ""
authors = [
    {name = "Максим",email = "maksim_goncharov_2006@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)"
]

[tool.poetry]
packages = [{include = "gateway", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
